safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-reporturi.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
  <body>
    <script src="../resources/securitypolicy-tests-base.js"></script>
    <script>
      if (document.securityPolicy.reportURIs.length === 0)
          log('PASS document.securityPolicy.reportURIs has length 0 when no policy exists.');
      else
          log('FAIL document.securityPolicy.reportURIs has length ' + document.securityPolicy.reportURIs.length + ' when no policy exists.');

      injectPolicy('report-uri http://example.com');

      if (document.securityPolicy.reportURIs.length === 1)
          log('PASS document.securityPolicy.reportURIs has length 1 when policy exists.');
      else
          log('FAIL document.securityPolicy.reportURIs has length ' + document.securityPolicy.reportURIs.length + ' when policy exists.');

      if (document.securityPolicy.reportURIs[0] === "http://example.com/")
          log('PASS document.securityPolicy.reportURIs[0] is correct.');
      else
          log('FAIL document.securityPolicy.reportURIs[0] is ' + document.securityPolicy.reportURIs[0] + ' (should be `http://example.com/`).');
    </script>
  </body>
</html>
