safari-600.7-branch/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-reporturi-expected.txt

PASS document.securityPolicy.reportURIs has length 0 when no policy exists.
PASS document.securityPolicy.reportURIs has length 1 when policy exists.
PASS document.securityPolicy.reportURIs[0] is correct.
